added python examples for sorting
